Video course

+++ controllers/admin/AdminVideoController.js
@@ -1,6 +1,19 @@
 const { Video, Course } = require("../../models");
 
 class VideoController {
+  static async readVideoDetail(req, res, next) {
+    try {
+      const { videoId } = req.params
+      const response = await Video.findOne({
+        where: { id: videoId },
+        attributes: ['id', 'name', 'videoUrl']
+      })
+      res.status(200).json(response)
+    } catch (err) {
+      next(err)
+    }
+  }
+
   static async addVideo(req, res, next) {
     try {
       const { courseId } = req.params;
@@ -29,17 +42,50 @@ class VideoController {
 
   static async deleteVideo(req, res, next) {
     try {
-      const { videoId } = req.params
+      const { videoId } = req.params;
+
+      const foundVideo = await Video.findByPk(videoId);
+
+      if (!foundVideo) {
+        throw { name: "VideoNotFound" };
+      }
 
       await Video.destroy({
         where: {
-          id: videoId
-        }
-      })
+          id: videoId,
+        },
+      });
 
-      res.status(200).json({message: "Video has been deleted"})
+      res.status(200).json({ message: "Video has been deleted" });
     } catch (err) {
-      next(err)
+      next(err);
+    }
+  }
+
+  // masih error
+  static async updateVideo(req, res, next) {
+    try {
+      const { videoId } = req.params;
+      const { name:inputName } = req.body;
+
+      const foundVideo = await Video.findByPk(videoId);
+
+      if (!foundVideo) {
+        throw { name: "VideoNotFound" };
+      }
+
+      const response = await Video.update(
+        { name: inputName },
+        {
+          where: { id: videoId },
+          returning: true,
+        }
+      );
+      console.log(response)
+
+      res.status(200).json(response);
+    } catch (err) {
+      next(err);
     }
   }
 }
